requirements.txt
albumentations==1.3.1
annotated-types==0.5.0
attrs==23.1.0
cattrs==23.1.2
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
docstring-to-markdown==0.12
efficientnet-pytorch==0.6.3
exceptiongroup==1.1.3
filelock==3.9.0
fonttools==4.42.1
idna==3.4
imageio==2.31.2
jedi==0.19.0
jedi-language-server==0.41.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==15.0.7
lsprotocol==2023.0.0a2
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
munch==4.0.0
networkx==3.0
numpy==1.24.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
parso==0.8.3
Pillow==9.3.0
pip==22.0.2
pretrainedmodels==0.7.4
pydantic==2.3.0
pydantic_core==2.6.3
pygls==1.0.2
pymap3d==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.28.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
segmentation-models-pytorch==0.2.0
setuptools==59.6.0
six==1.16.0
sympy==1.11.1
threadpoolctl==3.2.0
tifffile==2023.8.25
timm==0.4.12
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
triton==2.0.0
typeguard==3.0.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
wheel==0.37.1
